requirements.txt
annotated-types
anyio
asgiref
attrs
certifi
charset-normalizer
colorama
distro
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Faker
h11
httpcore
httpx
idna
inflection
jiter
jsonschema
jsonschema-specifications
pydantic
pydantic_core
PyJWT
PyMuPDF
PyMuPDFb
PyPDF2
python-dateutil
PyYAML
referencing
requests
rpds-py
six
sniffio
sqlparse
tqdm
typing_extensions
tzdata
uritemplate
urllib3
django-extensions
django-debug-toolbar
django-storages
boto3
python-dotenv
pillow
django-storages
boto3
zipfile36
imagequant
pypng
zopfli
docx
whitenoise